refactor: move SendMessageRequest record file
