--- src/main/java/com/project/dorumdorum/domain/chat/application/dto/SendMessageRequest.java
@@ -1,5 +1,0 @@
-package com.project.dorumdorum.domain.chat.application.dto;
-
-public record SendMessageRequest(
-        String initMessage
-) {}
