+++ src/main/java/com/project/dorumdorum/domain/chat/application/dto/request/SendMessageRequest.java
@@ -1,0 +1,5 @@
+package com.project.dorumdorum.domain.chat.application.dto.request;
+
+public record SendMessageRequest(
+        String initMessage
+) {}
