--- src/main/java/com/project/dorumdorum/domain/chat/ui/RejectMessageRequestController.java
@@ -1,4 +1,0 @@
-package com.project.dorumdorum.domain.chat.ui;
-
-public class RejectMessageRequestController {
-}
