+++ src/main/java/com/project/dorumdorum/domain/chat/ui/spec/SendMessageRequestApiSpec.java
@@ -1,6 +1,6 @@
 package com.project.dorumdorum.domain.chat.ui.spec;
 
-import com.project.dorumdorum.domain.chat.application.dto.SendMessageRequest;
+import com.project.dorumdorum.domain.chat.application.dto.request.SendMessageRequest;
 import com.project.dorumdorum.global.common.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
